Created types file and placed all types in it
