+++ src/pages/AlbumDetailPage.tsx
@@ -6,46 +6,9 @@ import AlbumDetail from "../components/AlbumDetail";
 import AlbumList from "../components/AlbumList";
 import PhotoList from "../components/PhotoList";
 import { useParams } from "react-router-dom";
+import { APIError, PhotoListData, AlbumDetailData, AlbumListData } from "../types";
 
-export type APIError = {
-  message: string;
-} | null;
-
-export interface ListData<T> {
-  next: boolean | null | string;
-  total: number;
-  results: T[]; //sustituye la t por lo que se ponga adentro de las cunas
-}
-
-export interface PhotoListItem {
-  id: string;
-  filename: string;
-  name: string;
-  date_and_time: string;
-}
-
-export interface AlbumListItem {
-  id: string;
-  name: string;
-  date: string;
-}
-
-export interface AlbumDetailItem {
-  id: string;
-  name: string;
-  date: string;
-  description: string;
-  breadcrumbs: {
-    id: string;
-    name: string;
-    date: string;
-  }[];
-}
-
-export type PhotoListData = ListData<PhotoListItem>;
-export type AlbumDetailData = AlbumDetailItem;
-export type AlbumListData = ListData<AlbumListItem>; //cambiar el nombre de esto porque es lo mismo que PhotoListData, un nombre que sirva para los dos;
-function fetchAlbum(albumId: string) {
+function fetchAlbum(albumId: string | undefined) {
   return axios({
     url: `${process.env.REACT_APP_API_BASE_URL}/api/v1/album/${albumId}`,
   }).then((res) => res.data);
@@ -78,7 +41,7 @@ function fetchPhotos(params: fetchPhotosParams) {
 function AlbumDetailPage() {
   const { albumId } = useParams();
   const albumQuery = useQuery<AlbumDetailData, APIError>(["album", albumId], () =>
-    fetchAlbum(albumId!),
+    fetchAlbum(albumId),
   );
   const childAlbumsQuery = useQuery<AlbumListData, APIError>(
     ["albums", { parent: albumId }],
