+++ src/pages/AlbumListPage.tsx
@@ -3,8 +3,8 @@ import React from "react";
 import { useQuery } from "react-query";
 import { Spinner } from "react-bootstrap";
 import AlbumList from "../components/AlbumList";
-import { APIError, AlbumListData } from "./AlbumDetailPage";
 
+import { APIError, AlbumListData } from "../types";
 function fetchAlbums() {
   return axios({
     url: `${process.env.REACT_APP_API_BASE_URL}/api/v1/albums`,
