--- src/pages/PhotoListData.tsx
@@ -1,3 +1,0 @@
-import { ListData, PhotoListItem } from "./AlbumDetailPage";
-
-export type PhotoListData = ListData<PhotoListItem>;
